Added ability to include transparency in output
The swatches in the swatches panel don't have an opacity setting, so
this update provides the ability to have a shape with a color from the
color scheme applied as well as an opacity value that then gets output
as a CSS value in rgba() notation.
Additionally, the demo file has been updated to reflect this.